Update on GOVERNANCE.md
TLDR: The reason CoreDNS had a special goverance structure different from
other CNCF project, was due to the fact that Miek is the original author of
CoreDNS and is still the single biggest (more than 50%) contributor of the project.

With Miek's intention to stepping down, a possible solution
to move forward is to change the  GOVERNANCE.md to have a
structure similiar to other CNCF projects (e.g., Envoy/Kubernetes).

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>